Add for variable debt token
